Remove all but active iframe on runner. Fixes #687
The old code was expecting a single irame to exist. Because it was only listening to the window.onload, we could render many times but only a single load event would fire. So code changed to search for all iframes, and remove all from the DOM except for 'sandbox.active'.