Fix #753. Capture runner.win.onerror and echo
Had to create proxyConsole._raw to allow for messages to *only* appear in the jsbin console, this is because I needed to show errors but not prevent the native error from throwing. The reason is: native errors have correct line numbers. If I had prevented the native error, the console.error in the native console would point to where I logged, not at the source of the error.

In addition, if I simply use console.error, the native console would show my manual call to console.error AND it would throw the native error, so the error appears twice on the native console - which looked weird.

The upshot is now the line numbers (in Chrome...though now I check Firefox, of course the numbers don't match at all) match up to the jsbin line number (because Chrome runs the script in it's own VM...or something, which matches the concept of splitting out the JS in jsbin). I think we may have to either remove the line numbers entirely or do a lot more work to get this accurate.